final changes for loop analysis
-new loop iteration counting system
--now counts all loops and in a single run
--this process now supports parallelism using an heuristic similar to
autopar's default (insert pragma in loops where nestedLevel == 0)
--besides code injection throught Clava it now requires additional files made
specially for this purpose
--results are stored in an object and later cached to prevent needless
recalculations

-final system for joinpoint count/analysis
--it now counts all jps of certain types found as descendants of the loop
--it also counts these totals taking into account inner loops and function
calls (supports multiple levels and mixing of the 2)
--from a subset of joinpoints some extra features are also extracted. these
values are from the joinpoint counts that support loops and function calls
--added aux functions for each jp type to help extract these features

-changes to functions that search and return loops

-added extra createXobj() functions

-adapted Polybench and NAS lara to support the new features and function calls

-NAS shows a problem when running all benchmarks when switching to SP. The
reason for this error is unclear as it adds wrong pragmas that don't seem
to come from other benchs or other parts of the same bench. Since when ran
alone it works fine this bug will be ignore for now as the desired results
can be extracted in the current form if the benchs are separated.

-added support to only calculate cache data and save it without timing the
benchmarks